chat-be/requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
djoser==2.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
MarkupSafe==2.0.1
msgpack==1.0.3
oauthlib==3.1.1
openapi-codec==1.3.2
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
python3-openid==3.2.0
pytz==2021.3
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
